Moved high score tracking to game class and add pom.xml for maven.
