+++ src/application/FreeHoldEm.java
@@ -61,14 +61,7 @@ public class FreeHoldEm {
 	 * Scanner used to get input from user.
 	 */
 	private static Scanner sc = new Scanner(System.in);
-	/**
-	 * Array to store the current high scores. It is initialized in assignHighScores function.
-	 */
-	private HighScore[] highScores;
-    /**
-     * Handles highScore file I/O.   
-     */
-	private HighScoreFile highScoreFile;
+
 	
 	private HandEvaluator he = new HandEvaluator();
 	
@@ -90,11 +83,6 @@ public class FreeHoldEm {
 	}
 	
 	public FreeHoldEm(int numPlayers) {
-		//Creates a new file or loads a existing file. 
-		highScoreFile = new HighScoreFile("high_scores.dat");
-		//Creates dummy highscores if there are no previous highscores.
-		highScoreFile.writeDummyHighScoresIfNecessary();
-		assignHighScores(highScoreFile);
 		players = new Player[numPlayers];
 		players[0] = new Player("human");
 		for (int i = 1; i < players.length; i++) {
@@ -108,22 +96,6 @@ public class FreeHoldEm {
 		//play();
 	}
 	
-	/**
-	 * Assigns loaded highscores into highScores array.
-	 * @param hf = HighScoreFile
-	 */
-	private void assignHighScores(HighScoreFile hf) {
-
-	  ArrayList<HighScore> list = hf.getHighScoreList();
-	  //Create a dynamic highscore array based on loaded highscores.
-	  highScores = new HighScore[list.size()];
-	  int i = 0;
-	  for (HighScore highScore : list) {
-	    highScores[i] = highScore;
-	    i++;
-	  }
-	}
-
 	public void dealHands() {
 		for (int i = 0; i < players.length; i++) {
 			if (players[i].checkFold()) {
@@ -211,73 +183,7 @@ public class FreeHoldEm {
 		}			
 	}
 	
-	private void play() {
-		//Deck deck = new Deck();
-		int j = 0;
-		//Deal player hands
-		for (int i = 0; i < players.length; i++) {
-			if (players[i].checkFold()) {
-				System.out.printf("%s is out of chips!", players[i].getName());
-			}
-			else {
-				players[i].setCards(0, deck.getCard(j));
-				players[i].setCards(1, deck.getCard(j + 1));
-				j = j + 2; 
-				if (i == 0) {
-					System.out.println("Your hand: " +
-							players[i].toString()); 
-				} 
-				else {
-					System.out.println(players[i].getName() +
-							": " + players[i].toString()); 
-				}      
-			}
-		}
-		Card[] yourHand = players[0].getCards();
-		System.out.printf("\n\n");
-		takeBets();
-		//Burn card before flop
-		j++;
-		//Deal flop                
-		cardsOnTable[0] = deck.getCard(j); 
-		cardsOnTable[1] = deck.getCard(j + 1); 
-		cardsOnTable[2] = deck.getCard(j + 2);
-		System.out.printf("Your Cards: %s, %s\n", yourHand[0].toString(), 
-				yourHand[1].toString());
-		System.out.printf("Flop: %s, %s, %s\n", cardsOnTable[0], 
-				cardsOnTable[1], cardsOnTable[2]);
-		System.out.println("Cash in pot: " + pot);
-		round++;
-		takeBets();
-		System.out.printf("\n\n");
-		//Burn card before turn
-		j += 4;
-		cardsOnTable[3] = deck.getCard(j);
-		System.out.printf("Your Cards: %s, %s\n", yourHand[0].toString(), yourHand[1].toString());
-		System.out.printf("Turn: %s, %s, %s, %s\n", cardsOnTable[0], cardsOnTable[1], 
-				cardsOnTable[2], cardsOnTable[3]);
-		System.out.println("Cash in pot: " + pot);
-		round++;
-		takeBets();
-		System.out.printf("\n\n");
-		//Burn card before river
-		j += 2;                
-		cardsOnTable[4] = deck.getCard(j);
-		System.out.printf("Your Cards: %s, %s\n", yourHand[0].toString(), yourHand[1].toString());
-		System.out.printf("River: %s, %s, %s, %s, %s\n", cardsOnTable[0], cardsOnTable[1], 
-				cardsOnTable[2], cardsOnTable[3], cardsOnTable[4]);
-		System.out.println("Cash in pot: " + pot);
-		round++;
-		takeBets();
-		System.out.printf("\n\n");
-		//Score player hands
-		for (int i = 0; i < players.length; i++) {
-			if (!(players[i].checkFold()))
-				he.findBestHand(this, players[i]);
-		}
-		//Pick winner
-		//Ensure a folded player cannot be chosen as the initial winner
-		//TODO: Handle ties (split winnings?)
+	public void pickWinner() {
 		int firstNotFoldedPlayer = -1;
 		for (int i = 0; i < players.length; i++) {
 			if (!(players[i].checkFold())) {
@@ -287,7 +193,6 @@ public class FreeHoldEm {
 		}
 		if (firstNotFoldedPlayer < 0) {
 			System.out.println("Everyone folded!");
-			startNewGame();
 		}
 		Player winner = players[firstNotFoldedPlayer];
 		for (int i = firstNotFoldedPlayer + 1; i < players.length; i++) {
@@ -301,14 +206,9 @@ public class FreeHoldEm {
 		winner.setChips(winner.getChips() + pot);
 		System.out.printf("\n%s wins with a \n%s\nWinnings: %d\n", winner.getName(),
 				winner.getHand().toString(), pot);
-		startNewGame();
 	}
-    /**
-     * Called at the end of each hand. If a player has folded, but still has
-     * chips they are allowed back into the next hand. If a player has no chips
-     * their fold flag is set to prevent them from playing in future games. 
-     */
-	private void startNewGame() {
+	
+	public void newHand() {
 		pot = 0;
 		for (int i = 0; i < players.length; i++) {
 			if (players[i].getChips() > 0) {
@@ -318,23 +218,7 @@ public class FreeHoldEm {
 				players[i].fold();
 			}
 		}
-		System.out.print("Would you like to play another hand? ");
-		String ans = sc.next();
-		if ((ans.equals("yes") || ans.equals("Yes")) && players[0].getChips() > 0) {
-			round = State.FIRST;
-			play();	
-		}
-		else if (players[0].getChips() <= 0) {
-			System.out.println("You are out of chips!");
-			//TODO: high score stuff
-			endGame();
-		}
-		else {
-			//TODO: high score stuff
-			endGame();
-		}
 	}
-
 	/**
 	 * Prompts user for input and places user bet. Once the AI bets are determined
 	 * with determineAIBet they are placed as well. If a player or CPU bets 0 chips
@@ -403,42 +287,19 @@ public class FreeHoldEm {
 		int bet = 0;
 		return bet;
 	}
-	/**
-	 * Called from startNewGame() if player declines to play another round or is
-	 * out of chips. This method determines if the players score is a high score
-	 * and if it is the writeNewScore method is called. If the player has not
-	 * made a high score the high scores are displayed and the game exits.
-	 */
-	private void endGame() {
-		//As they have sorted to descending order.
-		if (players[0].getChips() > highScores[highScores.length-1].getScore()) {
-			writeNewScore();
-		}
-		else {
-			//display high scores
-			System.out.println(highScoreFile.toString());
-		}
-		System.exit(0);
-		
-	}
-	/**
-	 * This method creates a new HighScore for the player and inserts it into the 
-	 * proper position in the highScores array. The scores are then displayed and
-	 * written to the high_scores.dat file.
-	 */
-	private void writeNewScore() {
-		System.out.printf("High Score!\nEnter your name: ");
-		String name = sc.next();
-		HighScore newHighScore = new HighScore(name, players[0].getChips());
-		highScoreFile.addHighScore(newHighScore);
-		//Sorts HighScores to descending order
-		highScoreFile.sortHighScores();
-		//Write highScores into file.
-		highScoreFile.writePlayersIntoFile();
-		System.out.println(highScoreFile.toString());
-	}
+
 	
 	public Card[] getCardsOnTable() {
 	  return this.cardsOnTable;
 	}
+	
+	public int getPlayerScore() {
+		return players[0].getChips();
+	}
+	
+	public boolean bettingComplete() {
+		return true;
+	}
+	
+
 }
