+++ src/test/TestHighScoreFile.java
@@ -1,4 +1,4 @@
-package unitTests;
+package test;
 
 import static org.junit.Assert.*;
